refactor: break up message tasks per model extracted
- each message processor is responsible for producing a single model
- add tracing throughout tasks
