+++ tasks/messages/types.go
@@ -1,0 +1,8 @@
+package messages
+
+import "github.com/ipfs/go-cid"
+
+type MessageError struct {
+	Cid   cid.Cid
+	Error string
+}
